better but still bad
